refactor: split reducers and add selectors for global state
