+++ src/index.js
@@ -3,7 +3,8 @@ import ReactDOM from 'react-dom';
 import {createStore, applyMiddleware} from 'redux';
 import thunk from 'redux-thunk';
 import {Provider} from 'react-redux';
-import {reducer, loadPokemons} from './js/store/reducer.js';
+import reducer from './js/store/reducer.js';
+import {loadPokemons} from './js/store/data/reducer.js';
 import App from './js/components/containers/App/App.jsx';
 import {getCurrentPage} from './js/utils.js';
 import './styles/index.scss';
