+++ src/js/store/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  DATA: 'DATA',
+  APP: 'APP',
+};
+
+export default NameSpace;
