+++ src/js/store/reducer.js
@@ -1,128 +1,9 @@
-import axios from 'axios';
-import Pokemon from '../models/pokemon.js';
-import {POKEMONS_PER_PAGE, getCatchDate} from '../utils.js';
+import {combineReducers} from 'redux';
+import {reducer as data} from './data/reducer.js';
+import {reducer as app} from './app/reducer.js';
+import NameSpace from './name-space.js';
 
-const initialState = {
-  pokemons: [],
-  caughtPokemons: [],
-  activePokemon: null,
-	errorMessage: null,
-	isLoading: false,
-};
-
-const ActionType = {
-  LOAD_POKEMONS: 'LOAD_POKEMONS',
-  UPDATE_POKEMON_STATUS: 'UPDATE_POKEMON_STATUS',
-  CHANGE_ACTIVE_POKEMON: 'CHANGE_ACTIVE_POKEMON',
-	CHANGE_ERROR_MESSAGE: 'CHANGE_ERROR_MESSAGE',
-	CHANGE_LOADING_STATUS: 'CHANGE_LOADING_STATUS',
-};
-
-const ActionCreator = {
-  loadPokemons: pokemons => {
-    return {
-      type: ActionType.LOAD_POKEMONS,
-      payload: pokemons,
-    };
-  },
-
-  updatePokemonStatus: pokemon => {
-    return {
-      type: ActionType.UPDATE_POKEMON_STATUS,
-      payload: pokemon,
-    };
-  },
-
-  changeActivePokemon: pokemon => {
-    return {
-      type: ActionType.CHANGE_ACTIVE_POKEMON,
-      payload: pokemon,
-    };
-  },
-
-  changeErrorMessage: message => {
-    return {
-      type: ActionType.CHANGE_ERROR_MESSAGE,
-      payload: message,
-    };
-  },
-
-  changeLoadingStatus: isLoading => {
-    return {
-      type: ActionType.CHANGE_LOADING_STATUS,
-      payload: isLoading,
-    };
-  },
-};
-
-const loadPokemons = (page) => (dispatch, getState) => {
-	dispatch(ActionCreator.changeLoadingStatus(true));
-	dispatch(ActionCreator.changeErrorMessage(null));
-
-  return axios.get(`http://localhost:3000/pokemons?_page=${page}&_limit=${POKEMONS_PER_PAGE}`)
-    .then(response => {
-      return Pokemon.parsePokemons(response.data);
-    })
-    .then(newPokemons => {
-			dispatch(ActionCreator.changeLoadingStatus(false));
-      dispatch(ActionCreator.loadPokemons(getState().pokemons.concat(newPokemons)));
-    })
-    .catch(err => {
-			dispatch(ActionCreator.changeLoadingStatus(false));
-      dispatch(ActionCreator.changeErrorMessage('Failed to load pokemons. Please try again later.'));
-      throw err;
-    });
-};
-
-const updatePokemonStatus = (caughtPokemons, pokemon) => {
-  if (!pokemon.isCaught) {
-    pokemon.isCaught = true;
-    pokemon.catchDate = getCatchDate();
-    caughtPokemons.unshift(pokemon);
-
-    return caughtPokemons;
-  } else {
-    pokemon.isCaught = false;
-    pokemon.catchDate = null;
-    
-    return caughtPokemons.filter(it => it.id !== pokemon.id);
-  }
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.LOAD_POKEMONS:
-      return {
-        ...state,
-        pokemons: action.payload,
-      };
-
-    case ActionType.UPDATE_POKEMON_STATUS:
-      return {
-        ...state,
-        caughtPokemons: updatePokemonStatus(state.caughtPokemons, action.payload),
-      };
-      
-    case ActionType.CHANGE_ACTIVE_POKEMON:
-      return {
-        ...state,
-        activePokemon: action.payload,
-      };
-      
-		case ActionType.CHANGE_ERROR_MESSAGE:
-			return {
-				...state,
-				errorMessage: action.payload,
-			};
-      
-		case ActionType.CHANGE_LOADING_STATUS:
-			return {
-				...state,
-				isLoading: action.payload,
-			};
-  }
-
-  return state;
-};
-
-export {reducer, loadPokemons, ActionCreator};
+export default combineReducers({
+  [NameSpace.DATA]: data,
+  [NameSpace.APP]: app,
+});
